requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
docstring_parser==0.16
exceptiongroup==1.2.1
executing==2.0.1
Flask==3.0.3
frozenlist==1.4.1
genai==2.1.0
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.132.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.49.0
google-cloud-bigquery==3.24.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.6.0
google-generativeai==0.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
ipykernel==6.29.4
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.2.3
langchain-core==0.2.5
langchain-google-genai==1.0.6
# langchain-google-genai
langchain-text-splitters==0.2.1
langsmith==0.1.75
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
multidict==6.0.5
nest-asyncio==1.6.0
numpy==1.26.4
openai==0.27.10
orjson==3.10.3
packaging==23.2
parso==0.8.4
platformdirs==4.2.2
prompt_toolkit==3.0.46
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.3
pydantic_core==2.18.4
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
rsa==4.9
shapely==2.0.4
six==1.16.0
SQLAlchemy==2.0.30
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.3.0
tiktoken==0.3.3
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.1
vertexai==1.49.0
wcwidth==0.2.13
Werkzeug==3.0.3
yarl==1.9.4
